Include locale redirect modules in pages controller
Now that we do not have to take care of the public child redirect anymore we can move the rather simple logic into the controller again.
